requirements.txt
requests
python-dotenv
rumps
streamlit
